+++ CHANGES.md
@@ -1,3 +1,5 @@
+- Replaced occurences of `is_plugin_active` by `class_exists`
+- Added reset link to plugins page
 - Added function remove_client_caps
 - Improved flexibility by adding filters
 - Reorganized code